### 2018-11-28 #1103 @whedon
[PRE REVIEW]: vcftoolz: a Python package for comparing and evaluating Variant Call Format files
**Submitting author:** @stevendavis (<a href="http://orcid.org/0000-0002-5581-1922">Steve Davis</a>)
**Repository:** <a href="https://github.com/CFSAN-Biostatistics/vcftoolz" target ="_blank">https://github.com/CFSAN-Biostatistics/vcftoolz</a>
**Version:** 1.0.0
