### 2019-01-17 #1144 @richyanicky
The software functioned as stated. It ran without error within the bounds of the description provided.
There are some issues that I came across that would be a benefit to our use.
1) Read gz format directly. I could not read our gz vcf files.
2) Files > 1M bp are too large to process and often caused our server to run out of memory.
3) The Venn diagrams are great, however it used the long file names and they could not be read in the output. A option would be to provide plot labels.
