### 2019-01-21 #1144 @mgymrek
Thanks @richyanicky for the review.

@stevendavis, could you respond to the reviewer comments, and in particular comment on the size of files and the compute requirements? Is this meant to process genome-wide VCFs? In which case it would be important to handle gzipped files. On the other hand if it is meant to handle only small files it is worth making that clear.
