### 2019-01-28 #1144 @stevendavis
@richyanicky I'm back to work after the government shutdown.  Thank you for the review comments.
1. Supporting .gz files is a good idea.
2. We use the software to process bacterial variants. In our work, we usually have dozens of samples in a multi-VCF file with 10,000 or fewer variant positions and our VCF files include only the variant positions.  Could you clarify, in your testing, do you have more than 1 million variant positions?  How many samples per VCF file?
3. An option to provide plot labels is a good idea.

@mgymrek Do these issues preclude publishing the manuscript, or would it be acceptable to address the issues in a future software release?
