### 2019-02-01 #1144 @mgymrek
@stevendavis: The changes are not strictly necessary, however I do have the following comments about @richyanicky's points

1. It looks like the package uses PyVCF to parse VCFs. I believe reading gzipped vcfs should be supported (e.g. something like `vcf.Reader(open('file.vcf.gz', 'rb'))` has worked for me)

2. Can you clarify in the paper the data sizes you are targeting. I know it mentions pathogens, but it would be helpful for a user to know it is meant for smaller variant files (vs. say human whole genome) by explicitly saying it is for on the order of thousands of variants.

3. Not necessary, but would be nice.

Also in the paper should the sentence: "The artifacts produced by vcftools are not available from other tools." say `vcftoolz` instead of `vcftools`?
