### 2019-02-05 ac28cbf
Support reading gz compressed vcf files.
