### 2019-02-06 7448562
Update version and history for v1.1.0.
