+++ setup.py
@@ -26,7 +26,7 @@ test_requirements = [
 
 setup(
     name='vcftoolz',
-    version='1.0.0',
+    version='1.1.0',
     description="Tools for working with Variant Call Format files.",
     long_description=readme + '\n\n' + history,
     author="Steve Davis",
