+++ HISTORY.rst
@@ -3,6 +3,11 @@
 History
 =======
 
+1.1.0 (2019-02-06)
+---------------------
+* Support reading gzip compressed vcf files.
+
+
 1.0.0 (2018-11-20)
 ---------------------
 
