### 2019-02-06 d4cce49
Update manuscript. Fix typo. Clarify expected VCF file sizes.
