### 2019-02-06 #1144 @stevendavis
@mgymrek The vcftoolz software is now updated to support reading gzipped vcf files.  Thanks for the tip.  I've bumped the version number to 1.1.0.  I have not yet pushed the release to PyPi, but will do so soon.

The manuscript is also updated to clarify the expected number of variants, and fixed the typo.  Good catch!

Do I need to do anything to trigger PDF file creation?
