### 2019-02-15 #1144 @mgymrek
@stevendavis this looks good. Can you add DOIs for the two paper references?
