+++ paper/paper.bib
@@ -4,7 +4,8 @@
   journal={BioRxiv},
   pages={023754},
   year={2015},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi={10.1101/023754}
 }
 
 @online{BCFtools,
@@ -31,5 +32,6 @@
   number={15},
   pages={2156--2158},
   year={2011},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btr330}
 }
