### 2019-02-21 #1144 @stevendavis
@mgmrek Thank you for the feedback.  The DOIs are added for the two paper references.  Let me know if any other changes are needed.
