### 2019-03-05 #1144 @stevendavis
The Zenodo DOI for the vcftoolz software is:
10.5281/zenodo.2583998

The version is:
v1.1.0
