### 2019-03-18 3028a2d
Replace None with '.' when printing call data.
