+++ HISTORY.rst
@@ -3,6 +3,10 @@
 History
 =======
 
+Unreleased
+---------------------
+* Replace None with '.' when printing call data.
+
 1.1.0 (2019-02-06)
 ---------------------
 * Support reading gzip compressed vcf files.
