### 2019-03-26 #1144 @stevendavis
@richyanicky @mgymrek @arfon Awesome! Thank you everyone for the comments and suggestions and for keeping my first JOSS submission running along smoothly.
